Trace options added for soundness (still buggy).

+++ src/dk/aau/cs/verification/Stats.java
@@ -51,7 +51,7 @@ public class Stats {
 		return stored;
 	}
 	
-	public long minimumExecutionTime(){
+	public int minimumExecutionTime(){
 		return minExecutionTime;
 	}
